fix: resolve 13 code review findings (2 critical, 4 security, 4 arch, 3 enhance)
Critical:
- C1: EthereumSignature.v changed from u8 to u64, fixing EIP-155
  signing for all L2s/sidechains with chain_id > 110 (Polygon, Arbitrum,
  Base, Optimism, etc.). Added recovery_bit() and to_bytes_eip155().
- C2: keypair_bytes() no longer leaks private key into plain Vec;
  copies directly from Zeroizing wrapper via extend_from_slice.

Security:
- S1: All 7 signers' generate() now use secure_random() instead of
  OsRng, respecting set_custom_rng() in TEE environments.
- S2: ct_hex_val is now fully branchless, returning (value, mask) tuple.
- S3: ct_hex_decode processes full input on odd-length (no early return).
- S4: XRP + NEO validate_address use subtle::ConstantTimeEq for checksum.

Architecture:
- A1: BLS re-exported from ethereum::bls (Ethereum consensus primitive).
- A2: Display impl added for XrpSignature, BlsSignature, NeoSignature.
- A3: Mnemonic from_phrase word lookup changed to binary search.
- A4: bip85::entropy_to_mnemonic delegates to Mnemonic::from_entropy.

Enhancements:
- E1: #[must_use] on all 7 signature types.
- E2: PartialEq/Eq derived for all 7 signature types.
- E3: to_seed salt String zeroized after use.

+++ src/ethereum/transaction.rs
@@ -193,7 +193,7 @@ impl EIP2930Transaction {
         items.extend_from_slice(&rlp::encode_u128(self.value));
         items.extend_from_slice(&rlp::encode_bytes(&self.data));
         items.extend_from_slice(&rlp::encode_access_list(&self.access_list));
-        items.extend_from_slice(&rlp::encode_u64(y_parity as u64));
+        items.extend_from_slice(&rlp::encode_u64(y_parity));
         items.extend_from_slice(&rlp::encode_bytes(&strip_leading_zeros(&sig.r)));
         items.extend_from_slice(&rlp::encode_bytes(&strip_leading_zeros(&sig.s)));
 
@@ -267,7 +267,7 @@ impl EIP1559Transaction {
         items.extend_from_slice(&rlp::encode_u128(self.value));
         items.extend_from_slice(&rlp::encode_bytes(&self.data));
         items.extend_from_slice(&rlp::encode_access_list(&self.access_list));
-        items.extend_from_slice(&rlp::encode_u64(y_parity as u64));
+        items.extend_from_slice(&rlp::encode_u64(y_parity));
         items.extend_from_slice(&rlp::encode_bytes(&strip_leading_zeros(&sig.r)));
         items.extend_from_slice(&rlp::encode_bytes(&strip_leading_zeros(&sig.s)));
 
@@ -359,7 +359,7 @@ impl EIP4844Transaction {
             hash_items.extend_from_slice(&rlp::encode_bytes(h));
         }
         items.extend_from_slice(&rlp::encode_list(&hash_items));
-        items.extend_from_slice(&rlp::encode_u64(y_parity as u64));
+        items.extend_from_slice(&rlp::encode_u64(y_parity));
         items.extend_from_slice(&rlp::encode_bytes(&strip_leading_zeros(&sig.r)));
         items.extend_from_slice(&rlp::encode_bytes(&strip_leading_zeros(&sig.s)));
